feat: message 모델 변경
